Tiny bug fixes

+++ drjava/src/edu/rice/cs/drjava/project/ProjectProfile.java
@@ -353,10 +353,10 @@ public class ProjectProfile implements ProjectFileIR {
         xc.set(".timestamp", s, f, true);
         String pkg = df.getPackage();
         xc.set(".package", (pkg != null)?pkg:"", f, true);
-        xc.set("select.from",   String.valueOf((pSel != null)?pSel.first():0),  f, true);
-        xc.set("select.to",     String.valueOf((pSel != null)?pSel.second():0), f, true);
-        xc.set("scroll.column", String.valueOf((pScr != null)?pScr.first():0),  f, true);
-        xc.set("scroll.row",    String.valueOf((pScr != null)?pScr.second():0), f, true);
+        xc.set("select.from",   String.valueOf((pSel != null) ? pSel.first()  : (Integer) 0), f, true);
+        xc.set("select.to",     String.valueOf((pSel != null) ? pSel.second() : (Integer) 0), f, true);
+        xc.set("scroll.column", String.valueOf((pScr != null) ? pScr.first()  : (Integer) 0), f, true);
+        xc.set("scroll.row",    String.valueOf((pScr != null) ? pScr.second() : (Integer) 0), f, true);
         if (df==active) xc.set(".active", "true", f, true);
       }
     }
@@ -383,11 +383,11 @@ public class ProjectProfile implements ProjectFileIR {
         xc.set(".name", path, f, true);
         xc.set(".timestamp", s, f, true);
         String pkg = df.getPackage();
-        xc.set(".package", (pkg != null)?pkg:"", f, true);
-        xc.set("select.from",   String.valueOf((pSel != null)?pSel.first():0),  f, true);
-        xc.set("select.to",     String.valueOf((pSel != null)?pSel.second():0), f, true);
-        xc.set("scroll.column", String.valueOf((pScr != null)?pScr.first():0),  f, true);
-        xc.set("scroll.row",    String.valueOf((pScr != null)?pScr.second():0), f, true);
+        xc.set(".package", (pkg != null) ? pkg:"", f, true);
+        xc.set("select.from",   String.valueOf((pSel != null) ? pSel.first()  : (Integer) 0), f, true);
+        xc.set("select.to",     String.valueOf((pSel != null) ? pSel.second() : (Integer) 0), f, true);
+        xc.set("scroll.column", String.valueOf((pScr != null) ? pScr.first()  : (Integer) 0), f, true);
+        xc.set("scroll.row",    String.valueOf((pScr != null) ? pScr.second() : (Integer) 0), f, true);
         if (df==active) { xc.set(".active", "true", f, true);
         }
       }
@@ -416,11 +416,11 @@ public class ProjectProfile implements ProjectFileIR {
         xc.set(".name", path, f, true);
         xc.set(".timestamp", s, f, true);
         String pkg = df.getPackage();
-        xc.set(".package", (pkg != null)?pkg:"", f, true);
-        xc.set("select.from",   String.valueOf((pSel != null)?pSel.first():0),  f, true);
-        xc.set("select.to",     String.valueOf((pSel != null)?pSel.second():0), f, true);
-        xc.set("scroll.column", String.valueOf((pScr != null)?pScr.first():0),  f, true);
-        xc.set("scroll.row",    String.valueOf((pScr != null)?pScr.second():0), f, true);
+        xc.set(".package", (pkg != null) ? pkg : "", f, true);
+        xc.set("select.from",   String.valueOf((pSel != null) ? pSel.first():(Integer)  0), f, true);
+        xc.set("select.to",     String.valueOf((pSel != null) ? pSel.second():(Integer) 0), f, true);
+        xc.set("scroll.column", String.valueOf((pScr != null) ? pScr.first():(Integer)  0), f, true);
+        xc.set("scroll.row",    String.valueOf((pScr != null) ? pScr.second():(Integer) 0), f, true);
         if (df==active) { xc.set(".active", "true", f, true);
         }
       }
@@ -436,7 +436,7 @@ public class ProjectProfile implements ProjectFileIR {
     xc.createNode("drjava/project/classpath");
     if (!_classPathFiles.isEmpty()) {
       for(AbsRelFile cp: _classPathFiles) {
-        path = cp.keepAbsolute()?cp.getAbsolutePath():FileOps.stringMakeRelativeTo(cp, _projectRoot);
+        path = cp.keepAbsolute() ? cp.getAbsolutePath():FileOps.stringMakeRelativeTo(cp, _projectRoot);
         path = replace(path, File.separator, "/");
         Node f = xc.createNode("drjava/project/classpath/file", null, false);
         xc.set(".name", path, f, true);
